New Page: Snippets Cheat Sheet
Focus to provide a quick reference on how to do general stuff in
HaxeFlixel, used in every project and all the time while coding one.
Not made for beginners, but can also be useful to them to see how
things are done, but the purpose isn't to give much explanation here
(that's what specific pages).

Needs a revision.