Update 00-community.html.md
We're an active, welcoming, and friendly community on Discord. The HF Discord channel isn't just a clone of the Slack server. It's less formal, and we're happy to talk about game development in general. We're not limited to discussions about the engine. It's also a place to show your project and get feedback. I think it complements the Slack server rather nicely so it would be a nice addition to the community.