Update 03-2-creating-a-new-project.html.md (#209)
I did my best to follow the tutorial, but it doesn't seem to have the fully qualified path seen here http://forum.haxeflixel.com/topic/544/type-not-found-flxtext/4, or instructions to import flixel.text.FlxText , so I figured I'd try to fix it myself.

Gama11's answer was written after the latest change to this file, so it's possible I just missed it. I did ctrl+f and search for the import, but the text doesn't seem to be on the page anywhere, and the template generated by my flixel-tools (up to date) doesn't have the import.

I originally had a small blurb trying to explain what FlxText is, but figured the less I changed things the better.

My apologies if I missed anything, and thanks for all you guys do!