Specify FlxColor package in 04-3-groundwork (#249)
Specify which package the FlxColor package lives in.

Every other example in the docs lists the import statement when adding a new package. For this example to docs were saying (Make sure that an import is created for FlxColor, and if it isn't, add it), which is far more confusing, and doesn't follow the example set by all other examples.. I for example attempted to use Flx.FlxColor at first, and was surprised to find it in Flx.util.FlxColor.